+++ lib/screens/provaider_screen.dart
@@ -3,6 +3,8 @@ import 'package:warrenapi/repository/cryptodata_repository_provider.dart';
 import 'package:warrenapi/usecase/usecase.dart';
 import 'package:warrenapi/usecase/viewdata.dart';
 
+import '../usecase/viewdata_chart.dart';
+
 final dataCryptoUseCase = Provider((ref) {
   return DataCryptoUseCase(repository: ref.read(dataCryptoRepositoryProvider));
 });
@@ -12,6 +14,7 @@ final dataCryptoProvider = FutureProvider.autoDispose
   return ref.read(dataCryptoUseCase).execute();
 });
 
-final dataChartProvider = FutureProvider<List<DataChartsViewData>>((ref) async {
+final dataChartProvider =
+    FutureProvider.autoDispose<List<DataChartsViewData>>((ref) async {
   return ref.read(dataCryptoUseCase).executechart();
 });
