+++ lib/usecase/mapper.dart
@@ -1,5 +1,6 @@
 import 'package:warrenapi/usecase/viewdata.dart';
 import '../models/api model/crypto_full_data_response.dart';
+import 'viewdata_chart.dart';
 
 extension CryptoMapper on FullDataResponse {
   List<DataCryptoViewData> toViewData() {
