+++ lib/usecase/usecase.dart
@@ -1,6 +1,6 @@
 import 'package:warrenapi/usecase/mapper.dart';
 import 'package:warrenapi/usecase/viewdata.dart';
-
+import 'package:warrenapi/usecase/viewdata_chart.dart';
 import '../repository/cryptodata_repository.dart';
 
 class DataCryptoUseCase {
