+++ lib/usecase/viewdata.dart
@@ -12,9 +12,3 @@ class DataCryptoViewData {
       required this.slug,
       required this.symbol});
 }
-
-class DataChartsViewData {
-  final List<List<double>> btc_timeseries;
-
-  DataChartsViewData({required this.btc_timeseries});
-}
