+++ lib/usecase/viewdata_chart.dart
@@ -1,0 +1,5 @@
+class DataChartsViewData {
+  final List<List<double>> btc_timeseries;
+
+  DataChartsViewData({required this.btc_timeseries});
+}
